Reverting Merge and Daily Data push
This reverts commit d574a82392b07da74d47b849501c49c6a1a5f3a9, reversing
changes made to e6be57c1b9407a8c05213536f147f9862831d26c.

Revert "-Think I fixed the Daily Data issue"

This reverts commit e6be57c1b9407a8c05213536f147f9862831d26c.
