+++ .idea/Astro2.iml
@@ -25,7 +25,6 @@
       <sourceFolder url="file://$MODULE_DIR$/V1 AstroSolver/src/transforms/GroundBasedTransforms/warner_aux" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/V1 AstroSolver/src/transforms/SpaceBasedTransforms" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/V1 AstroSolver/src/transforms/TrackRateModeTransforms" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/V1 AstroSolver/src/functions" isTestSource="false" />
     </content>
     <orderEntry type="jdk" jdkName="Python 3.8 (Astro2-updated)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
