+++ V1 AstroSolver/src/pinpointsolving/pinpoint.py
@@ -4,7 +4,6 @@ import os
 
 import win32com
 import win32com.client
-
 import AstroFunctions as astro
 
 
@@ -99,8 +98,3 @@ def pinpoint_solve(inbox, catloc, max_mag, sigma, catexp, match_residual,
                     except:
                         print("Could Not Solve")
                         continue
-
-
-def pinpoint_init():
-    f = win32com.client.Dispatch("Pinpoint.plate")
-    return f
