+++ V1 AstroSolver/src/transforms/GroundBasedTransforms/warner_aux/auxilary_phot_warner_functions.py
@@ -7,12 +7,17 @@ Created on Fri Apr 29 14:31:56 2022
 This file contains specific auxillary funcitons which correspond to the warner transformations
 """
 import os
+
+
+import numpy
+import numpy as np
+
+from astropy.table import Table, hstack
+
 from math import sqrt
 
-import matplotlib.cm as cm
-import numpy as np
-from astropy.table import Table, hstack
 from matplotlib import pyplot as plt
+
 
 import AstroFunctions as astro
 
@@ -91,8 +96,9 @@ def apply_transforms_Warner(stars_table,
         plt.close()
 
 
-def exoatmospheric_mags_warner(stars_table,
-                               extinction_table_warner,
+
+def exoatmospheric_mags_Warner(stars_table,
+                               extinction_table_Warner,
                                different_filter_list):
     # m_0 = m - k'_f * X - k''_fCI * X * CI
     nan_array = np.empty(len(stars_table))
@@ -174,9 +180,9 @@ def exoatmospheric_mags_warner(stars_table,
             colour_indices = astro.get_all_colour_indices(different_filter)
             x_column_name = f"X_{different_filter}"
             for colour_index in colour_indices:
-                mask = ((extinction_table_warner['filter'] == different_filter) & (
-                        extinction_table_warner['CI'] == colour_index))
-                row_of_extinctions = extinction_table_warner[mask]
+                mask = ((extinction_table_Warner['filter'] == different_filter) & (
+                    extinction_table_Warner['CI'] == colour_index))
+                row_of_extinctions = extinction_table_Warner[mask]
                 # if len(row_of_transforms) == 0 and instr_filter == 'v':
                 #     instr_filter = 'g'
                 #     mask = ((gb_final_transforms['filter'] == instr_filter) & (gb_final_transforms['CI'] == colour_index))
@@ -668,7 +674,7 @@ def exoatmospheric_mags_verify_Warner(stars_table,
 
 
 
-def second_order_extinction_calc_warner(slopes_table,
+def second_order_extinction_calc_Warner(slopes_table,
                                         different_filter_list,
                                         save_plots, **kwargs):
     '''
@@ -830,209 +836,3 @@ def second_order_extinction_calc_warner(slopes_table,
             k_prime_sigma_column
         ])
     return extinction_table_Warner
-def calculate_slopes_Warner(stars_table, different_filter_list, save_plots, **kwargs):
-    '''
-    Calculates the slope of Each Stars Magntitude with Time and plots it
-
-    Parameters
-    ----------
-    stars_table :  astropy.table.table.Table
-        Table containing the mean of the important information for each star.
-        Has columns:
-            Field : string
-                Unique identifier of the star field that the reference star
-                is in (e.g. Landolt field "108").
-            Name : string
-                Name/unique identifier of the reference star.
-            V : numpy.float64
-                Apparent V magnitude from the reference file.
-            (B-V) : numpy.float64
-                Apparent B-V colour index from the reference file.
-            (U-B) : numpy.float64
-                Apparent U-B colour index from the reference file.
-            (V-R) : numpy.float64
-                Apparent V-R colour index from the reference file.
-            (V-I) : numpy.float64
-                Apparent V-I colour index from the reference file.
-            V_sigma : numpy.float64
-                Standard deviation of the apparent V magnitude from
-                the reference file.
-            <filter> : numpy.float64
-                Mean instrumental magnitude of all detections of the star in
-                <filter>. There is a different column for
-                each different filter used across the images.
-            <filter>_sigma : numpy.float64
-                Standard deviation of the instrumental magnitudes of all
-                detections of the star in <filter>.
-                There is a different column for each different filter
-                used across the images.
-            X_<filter> : numpy.float64
-                Mean airmass of all detections of the star in <filter>.
-                There is a different column for each different
-                filter used across the images. Only output if ground_based
-                is True.
-            X_<filter>_sigma : numpy.float64
-                Standard deviation of the airmasses of all detections of
-                the star in <filter>. There is a different
-                column for each different filter used across the images.
-                Only output if ground_based is True.
-
-    different_filter_list : List
-        Different Filter List
-    save_plots : Boolean
-        True= Save Plots
-    **kwargs : TYPE
-        DESCRIPTION.
-
-    Returns
-    -------
-    slopes_table : Table
-        DESCRIPTION.
-
-    '''
-
-    stars_for_second_order_extinction, multiple_stars = astro.get_stars_with_multiple_observations(
-        stars_table)
-    slope_filters = [
-        f"slope_{different_filter}" for different_filter in different_filter_list]
-    intercept_filters = [
-        f"intercept_{different_filter}" for different_filter in different_filter_list]
-    slope_filters_sigma = [
-        f"slope_{different_filter}_sigma" for different_filter in different_filter_list]
-    intercept_filters_sigma = [
-        f"intercept_{different_filter}_sigma" for different_filter in different_filter_list]
-    nan_array = np.empty(len(multiple_stars))
-    nan_array.fill(np.nan)
-    data_filter_table = [
-        nan_array for different_filter in different_filter_list]
-    star_index_columns = [
-        'Field',
-        'Name',
-        'V_ref',
-        'B-V',
-        'U-B',
-        'V-R',
-        'V-I',
-        'V_sigma',
-        'e_B-V',
-        'e_U-B',
-        'e_V-R',
-        'e_V-I'
-    ]
-    star_index_table = Table(
-        names=star_index_columns,
-        data=[
-            np.empty(len(multiple_stars), dtype=object),
-            np.empty(len(multiple_stars), dtype=object),
-            nan_array,
-            nan_array,
-            nan_array,
-            nan_array,
-            nan_array,
-            nan_array,
-            nan_array,
-            nan_array,
-            nan_array,
-            nan_array
-        ]
-    )
-    slope_table = Table(names=slope_filters, data=data_filter_table)
-    intercept_table = Table(names=intercept_filters, data=data_filter_table)
-    slope_sigma_table = Table(
-        names=slope_filters_sigma, data=data_filter_table)
-    intercept_sigma_table = Table(
-        names=intercept_filters_sigma, data=data_filter_table)
-    slopes_table = hstack([star_index_table, slope_table,
-                          intercept_table, slope_sigma_table,
-                          intercept_sigma_table])
-    colors = cm.rainbow(np.linspace(0, 1, len(multiple_stars)))
-    for unique_filter in different_filter_list:
-        # current_filter = stars_table[unique_filter]
-        x_current_filter = f"X_{unique_filter}"
-        # unique_stars = table.unique(current_filter, keys='Name')
-        # colors = cm.rainbow(np.linspace(0, 1, len(multiple_stars)))
-        for i, unique_star in enumerate(multiple_stars):
-            # print(current_filter)
-            star_mask = stars_table['Name'] == unique_star
-            current_star = stars_table[star_mask]
-            # print(current_star[star_index_columns][0])
-            # print(current_star[star_index_columns])
-            for field in star_index_columns:
-                slopes_table[field][i] = current_star[field][0]
-            # slopes_table[star_index_columns][i] = list(current_star[star_index_columns][0])
-            # print(unique_filter)
-            # print(current_star)
-            x_nan_indices = np.isnan(current_star[x_current_filter])
-            y_nan_indices = np.isnan(current_star[unique_filter])
-            nan_indices = (x_nan_indices | y_nan_indices)
-            try:
-                X_plot = np.arange(
-                    start=min(current_star[x_current_filter][~np.isnan(
-                        current_star[x_current_filter])]) - 0.02,
-                    stop=max(current_star[x_current_filter][~np.isnan(
-                        current_star[x_current_filter])]) + 0.02,
-                    step=0.01)
-            except ValueError:
-                slopes_table[f"slope_{unique_filter}"][i] = np.nan
-                slopes_table[f"intercept_{unique_filter}"][i] = np.nan
-                slopes_table[f"slope_{unique_filter}_sigma"][i] = np.nan
-                slopes_table[f"intercept_{unique_filter}_sigma"][i] = np.nan
-                continue
-            # m, b = np.polyfit(current_star[x_current_filter], current_star[unique_filter], 1)
-            fit, or_fit, line_init = astro.init_linear_fitting(sigma=2.5)
-            try:
-                fitted_line, mask = or_fit(line_init, current_star[x_current_filter][~nan_indices],
-                                           current_star[unique_filter][~nan_indices])
-            except TypeError:
-                # print(current_star[x_current_filter])
-                # print(current_star[unique_filter])
-                slopes_table[f"slope_{unique_filter}"][i] = np.nan
-                slopes_table[f"intercept_{unique_filter}"][i] = np.nan
-                slopes_table[f"slope_{unique_filter}_sigma"][i] = np.nan
-                slopes_table[f"intercept_{unique_filter}_sigma"][i] = np.nan
-                continue
-            filtered_data = np.ma.masked_array(
-                current_star[unique_filter][~nan_indices], mask=mask)
-            m = fitted_line.slope.value
-            b = fitted_line.intercept.value
-            cov = fit.fit_info['param_cov']
-            try:
-                m_sigma = sqrt(cov[0][0])
-                b_sigma = sqrt(cov[1][1])
-            except TypeError:
-                m_sigma = np.nan
-                b_sigma = np.nan
-            plt.plot(current_star[x_current_filter],
-                     current_star[unique_filter], 'o', fillstyle='none',
-                     color=colors[i], label="Clipped Data")
-            plt.plot(current_star[x_current_filter][~nan_indices], filtered_data, 'o', color=colors[i],
-                     label="Fitted Data")
-            # plt.scatter(current_star[x_current_filter], current_star[unique_filter], color=colors[i], label=unique_star)
-            plt.plot(X_plot, m * X_plot + b, color=colors[i])
-            slopes_table[f"slope_{unique_filter}"][i] = m
-            slopes_table[f"intercept_{unique_filter}"][i] = b
-            slopes_table[f"slope_{unique_filter}_sigma"][i] = m_sigma
-            slopes_table[f"intercept_{unique_filter}_sigma"][i] = b_sigma
-        # plt.plot(current_filter['X'], current_filter['mag_instrumental'], 'o')
-        plt.xlabel('X')
-        plt.ylabel(unique_filter.lower())
-        plt.title("Slope of a Star's magnitude vs. airmass")
-        plt.ylim([min(stars_table[unique_filter][~np.isnan(stars_table[unique_filter])]) * 1.05,
-                  max(stars_table[unique_filter][~np.isnan(stars_table[unique_filter])]) * 0.95])
-        plt.gca().invert_yaxis()
-        # plt.title(unique_field)
-        # plt.legend()
-        if save_plots:
-            save_loc = f"{os.path.join(kwargs.get('save_loc'), f'Slopes{unique_filter}')}.png"
-            plt.savefig(save_loc)
-        plt.show()
-        plt.close()
-    # Column names to keep.
-    # 'Field','Name','V_ref','B-V','U-B','V-R','V-I','V_sigma','e_B-V','e_U-B','e_V-R','e_V-I'
-    # print(stars_for_second_order_extinction.columns)
-    # slopes_table.pprint(max_lines=-1, max_width=250)
-    return slopes_table
-
-
-
-
