+++ V1 AstroSolver/src/transforms/TrackRateModeTransforms/RunSCLightCurve.py
@@ -5,6 +5,9 @@ Created on Thu Jun  3 11:16:13 2021
 
 @author: jmwawrow
 """
+import AstroFunctions as astro
+from astropy.io import ascii
+import matplotlib
 import os
 
 # directory = 'D:\\Transfer to mac\\2021-03-10 - Calibrated\\Intelsat 10-02 Post Eclipse\\LIGHT\\B_lim'
@@ -12,8 +15,6 @@ import os
 # directory = r'C:\Users\jmwawrow\Documents\DRDC_Code\Intelsat 10-02\2021-04-21\Intelsat 10-02 ALL'
 # directory = r'C:\Users\jmwawrow\Documents\DRDC_Code\NEOSSat Observations\2016-072'
 # directory = r'F:\Intelsat 10-02\2020 11 30 - 2x2 - unprocessed\Intelsat 10-02\corrected_lights'
-import trm_auxillary_functions
-
 directory = r'C:\Users\mstew\Documents\School and Work\Winter 2022\Work\Astro2_Files\StJohnsTargets\2021_04_25_StJohns_Targets\2021_04_25_StJohns_Targets\2021-04-25 - unprocessed\Intelsat 10-02\LIGHT'
 
 # transforms_file = r'C:\Users\jmwawrow\Documents\DRDC_Code\Intelsat 10-02\2021-04-21\Solved Stars\Outputs\gb_final_transforms.csv'
@@ -47,16 +48,16 @@ plot_results = 0
 # gb_final_transforms = ascii.read(transforms_file)
 # directory = f"F:\\Intelsat 10-02\\{subfolder}\\Intelsat 10-02\\corrected_lights"
 save_loc = os.path.join(directory, 'Outputs_Warner_Test_TEMP')
-sat_dict, app_sat_dict, sats_table, uncertainty_table, sat_auxiliary_table = trm_auxillary_functions._main_sc_lightcurve(directory,
-                                                                                                                         gb_final_transforms=gb_final_transforms,
-                                                                                                                         temp_dir=temp_dir,
-                                                                                                                         save_loc=save_loc,
-                                                                                                                         file_suffix=file_suffix,
-                                                                                                                         ecct_cut=ecct_cut,
-                                                                                                                         max_distance_from_sat=max_distance_from_sat,
-                                                                                                                         size=size,
-                                                                                                                         max_num_nan=max_num_nan,
-                                                                                                                         plot_results=plot_results)
+sat_dict, app_sat_dict, sats_table, uncertainty_table, sat_auxiliary_table = astro._main_sc_lightcurve(directory,
+                                                                                                       gb_final_transforms=gb_final_transforms,
+                                                                                                       temp_dir=temp_dir,
+                                                                                                       save_loc=save_loc,
+                                                                                                       file_suffix=file_suffix,
+                                                                                                       ecct_cut=ecct_cut,
+                                                                                                       max_distance_from_sat=max_distance_from_sat,
+                                                                                                       size=size,
+                                                                                                       max_num_nan=max_num_nan,
+                                                                                                       plot_results=plot_results)
 
 # print(sat_dict)
 # sat_auxiliary_table.pprint_all()
